requirements.txt
numpy~=1.24.3
torchvision~=0.15.2
pillow~=9.4.0
yaml~=0.2.5
pyyaml~=6.0
yacs~=0.1.6
sympy~=1.11.1
lightning~=2.0.2
torchmetrics~=0.11.4
tqdm~=4.65.0
pandas~=1.5.3
pynverse~=0.1.4.6
pytorch~=2.0.1
matplotlib~=3.7.1
timm~=0.9.2